Events-Testing-20251222
couldn't fingure out how to fork a branch as my vscode client seemed to not know there was a repo. once I repopend it. this is a branch from main as of 20251222 with a zip file containing the 20251209 game with the new tests in for the lever+spikes, the exit door, the bookshelf and book.